Handle rotation in detection module
Before this commit, the detector was assuming the screen will never change. This is not the case, as after a rotation, the width and height are swapped.
This was leading inconsistencies and possible crashes when an activity change the orientation after a scenario was loaded.

Now, when we detect a rotation, the virtual display is destroyed and rebuilt with the correct display value, fixing all of the above.