Fix possible crash on capture selection
When part of the selection is made outside the capture, it might crash.
Due to the multiple resizing and moving, the final size of the intersected selector might have a delta of 0.01. Unfortunately, the method RectF::toRect() is using ceil as a rounding method, which will leads to one pixel outside the capture. This only pixel will cause a crash at the bitmap creation, because the size values are incorrect.